Move auth forms into seperate component folder
