+++ app/src/components/authentication/SignInForm.tsx
@@ -11,7 +11,7 @@ import GoogleIcon from '@mui/icons-material/Google';
 import {useFormik} from 'formik';
 import {object, string} from 'yup';
 
-import {PrimaryButton, SecondaryButton} from './Button';
+import {PrimaryButton, SecondaryButton} from '../common/Button';
 import {SignInRequest} from '../../services/auth';
 
 interface SignInFormProps {
