+++ app/src/views/AccountVerification.tsx
@@ -3,7 +3,7 @@ import {styled} from '@mui/system';
 import {Stack, Typography} from '@mui/material';
 
 import {VerificationRequest, useVerificationMutation} from '../services/auth';
-import {AccountVerificationForm} from '../components/common/AccountVerificationForm';
+import {AccountVerificationForm} from '../components/authentication/AccountVerificationForm';
 
 export const AccountVerification = () => {
   const [verify] = useVerificationMutation();
